pyproject.toml
[project]
name = "plat-cam-py-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "fast-alpr[onnx]>=0.3.0",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "python-fasthtml>=0.12.33",
]
